Начинаем изучать Fragment. Fragment это как бы отдельный модуль который запускается на Activity при этом у него свой собственный цикл жизни который зависит от цикла жизни Activity.
