+++ .idea/misc.xml
@@ -3,9 +3,9 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/Users/react/AndroidStudioProjects/DrawerLayout/app/src/main/res/layout/activity_main.xml" value="0.264" />
-        <entry key="..\:/Users/react/AndroidStudioProjects/DrawerLayout/app/src/main/res/layout/header.xml" value="0.264" />
-        <entry key="..\:/Users/react/AndroidStudioProjects/DrawerLayout/app/src/main/res/menu/nav_menu.xml" value="0.38333333333333336" />
+        <entry key="..\:/Users/react/AndroidStudioProjects/Fragment/app/src/main/res/layout/activity_main.xml" value="0.38333333333333336" />
+        <entry key="..\:/Users/react/AndroidStudioProjects/Fragment/app/src/main/res/layout/fragment_blank.xml" value="0.38333333333333336" />
+        <entry key="..\:/Users/react/AndroidStudioProjects/Fragment/app/src/main/res/layout/fragment_blank2.xml" value="0.38333333333333336" />
       </map>
     </option>
   </component>
