+++ app/src/androidTest/java/com/example/fragment/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.drawerlayout
+package com.example.fragment
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.drawerlayout", appContext.packageName)
+        assertEquals("com.example.fragment", appContext.packageName)
     }
 }
