--- app/src/main/res/menu/nav_menu.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:showIn="navigation_view">
-    <item android:title="Category 1">
-        <menu>
-            <item android:title="Item 1" android:id="@+id/item1" android:icon="@drawable/ic_android_black_24dp"/>
-            <item android:title="Item 2" android:id="@+id/item2" android:icon="@drawable/ic_android_black_24dp"/>
-            <item android:title="Item 3" android:id="@+id/item3" android:icon="@drawable/ic_android_black_24dp"/>
-            <item android:title="Item 4" android:id="@+id/item4" android:icon="@drawable/ic_android_black_24dp"/>
-        </menu>
-    </item>
-
-    <item android:title="Category 2">
-        <menu>
-            <item android:title="Item 11" android:id="@+id/item11" android:icon="@drawable/ic_android_black_24dp"/>
-            <item android:title="Item 21" android:id="@+id/item21" android:icon="@drawable/ic_android_black_24dp"/>
-            <item android:title="Item 31" android:id="@+id/item31" android:icon="@drawable/ic_android_black_24dp"/>
-            <item android:title="Item 41" android:id="@+id/item41" android:icon="@drawable/ic_android_black_24dp"/>
-        </menu>
-    </item>
-
-</menu>
