+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "DrawerLayout"
+rootProject.name = "Fragment"
 include ':app'
